Extract Opencast helpers
